bindings for in rectangle

+++ integration/duckdb/src/lance/duckdb/vector_functions.cc
@@ -113,34 +113,15 @@ std::vector<std::unique_ptr<::duckdb::CreateFunctionInfo>> GetVectorFunctions() 
                                ::duckdb::LogicalType::ANY,
                                nullptr,
                                L2DistanceBind));
-  //  l2_distance.AddFunction(
-  //      ::duckdb::ScalarFunction({::duckdb::LogicalType::LIST(::duckdb::LogicalType::DOUBLE),
-  //                                ::duckdb::LogicalType::LIST(::duckdb::LogicalType::DOUBLE)},
-  //                               ::duckdb::LogicalType::DOUBLE,
-  //                               L2Distance<double>));
   functions.emplace_back(std::make_unique<::duckdb::CreateScalarFunctionInfo>(l2_distance));
 
   ::duckdb::ScalarFunctionSet in_rectangle("in_rectangle");
-  //  in_rectangle.AddFunction(::duckdb::ScalarFunction(
-  //      {::duckdb::LogicalType::LIST(::duckdb::LogicalType::INTEGER),
-  //       ::duckdb::LogicalType::LIST(::duckdb::LogicalType::LIST(::duckdb::LogicalType::INTEGER))},
-  //      ::duckdb::LogicalType::BOOLEAN,
-  //      IsInRectangle));
-  //  in_rectangle.AddFunction(::duckdb::ScalarFunction(
-  //      {::duckdb::LogicalType::LIST(::duckdb::LogicalType::BIGINT),
-  //       ::duckdb::LogicalType::LIST(::duckdb::LogicalType::LIST(::duckdb::LogicalType::BIGINT))},
-  //      ::duckdb::LogicalType::BOOLEAN,
-  //      IsInRectangle));
+  /// All upcast to double
   in_rectangle.AddFunction(::duckdb::ScalarFunction(
-      {::duckdb::LogicalType::LIST(::duckdb::LogicalType::FLOAT),
-       ::duckdb::LogicalType::LIST(::duckdb::LogicalType::LIST(::duckdb::LogicalType::FLOAT))},
+      {::duckdb::LogicalType::LIST(::duckdb::LogicalType::DOUBLE),
+       ::duckdb::LogicalType::LIST(::duckdb::LogicalType::LIST(::duckdb::LogicalType::DOUBLE))},
       ::duckdb::LogicalType::BOOLEAN,
       IsInRectangle));
-  //  in_rectangle.AddFunction(::duckdb::ScalarFunction(
-  //      {::duckdb::LogicalType::LIST(::duckdb::LogicalType::DOUBLE),
-  //       ::duckdb::LogicalType::LIST(::duckdb::LogicalType::LIST(::duckdb::LogicalType::DOUBLE))},
-  //      ::duckdb::LogicalType::BOOLEAN,
-  //      IsInRectangle));
   functions.emplace_back(std::make_unique<::duckdb::CreateScalarFunctionInfo>(in_rectangle));
 
   return functions;
